[Core] Introduce NixlStorageBackend dynamic mode (#2024)
* [Core] Seperate NixlStorageBackend classes

This commit prepares the backend for integration of
DynamicStorageBackend, implemented in the next commit.

Signed-off-by: Tomer Shmilovich <tshmilovich@nvidia.com>

* [Core] Introduce NixlStorageBackend dynamic mode

This commit incorporates PR #1939 into NixlStorageBackend, under the
name of NixlDynamicStorageBackend, in order to prevent having 2 nixl
storage backends as well as lessen code duplication.

In order to use this new dynamic storage mode, which creates object keys
when needed instead of a limited pre-allocated amount,
extra_config.nixl_pool_size should be set to 0.

This mode is currently only supported for nixl OBJ backend.

Signed-off-by: Tomer Shmilovich <tshmilovich@nvidia.com>

* [Docs] Update docs

Signed-off-by: Tomer Shmilovich <tshmilovich@nvidia.com>

* assert on save_unfull_chunk

Signed-off-by: Tomer Shmilovich <tshmilovich@nvidia.com>

* Generalize create_batched_storage_handler

Signed-off-by: Tomer Shmilovich <tshmilovich@nvidia.com>

* Fix obj ref_count leak

Signed-off-by: Tomer Shmilovich <tshmilovich@nvidia.com>

* Fix NIXL release order

Signed-off-by: Tomer Shmilovich <tshmilovich@nvidia.com>

* Generalize nixl_obj_exists

Signed-off-by: Tomer Shmilovich <tshmilovich@nvidia.com>

* Improve async usage

Signed-off-by: Tomer Shmilovich <tshmilovich@nvidia.com>

* Align get_blocking

Signed-off-by: Tomer Shmilovich <tshmilovich@nvidia.com>

---------

Signed-off-by: Tomer Shmilovich <tshmilovich@nvidia.com>